backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
stripe
python-dotenv
motor
pydantic